# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ONDC Integration API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Vaatun Technologies",
      "email": "vaibhav@vaatun.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://ondc-staging.vaatun.com",
      "description": "Staging environment"
    }
  ],
  "components": {
    "schemas": {
...
      "AckResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "object",
            "properties": {
              "ack": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACK",
                      "NACK"
                    ],
                    "example": "ACK"
                  }
                },
                "required": [
                  "status"
                ]
              }
            },
            "required": [
              "ack"
            ]
          },
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "message"
        ]
      },
      "BecknContext": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "example": "ONDC:FIS13"
          },
          "action": {
            "type": "string",
            "example": "on_search"
          },
          "bap_id": {
            "type": "string",
            "example": "ondc-staging.vaatun.com"
          },
          "bap_uri": {
            "type": "string",
            "example": "https://ondc-staging.vaatun.com/api/ondc"
          },
          "bpp_id": {
            "type": "string",
            "example": "ondc-mock-server.example.com"
          },
          "bpp_uri": {
            "type": "string",
            "example": "https://ondc-mock-server.example.com"
          },
          "transaction_id": {
            "type": "string",
            "example": "019abc12-3456-7890-abcd-ef1234567890"
          },
          "message_id": {
            "type": "string",
            "example": "019abc12-3456-7890-abcd-ef1234567891"
          },
          "timestamp": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "version": {
            "type": "string",
            "example": "2.0.1"
          }
        },
        "additionalProperties": {}
      },
      "OnSearchRequest": {
        "type": "object",
        "properties": {
          "context": {
            "$ref": "#/components/schemas/BecknContext"
          },
          "message": {
            "type": "object",
            "properties": {
              "catalog": {
                "description": "BPP catalog with providers and items"
              }
            },
            "additionalProperties": {}
          },
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "additionalProperties": {}
          }
        },
        "additionalProperties": {}
      },
      "OnSelectAckResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "object",
            "properties": {
              "ack": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACK",
                      "NACK"
                    ],
                    "example": "ACK"
                  }
                },
                "required": [
                  "status"
                ]
              }
            },
            "required": [
              "ack"
            ]
          },
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "message"
        ]
      },
      "OnSelectRequest": {
        "type": "object",
        "properties": {
          "context": {
            "type": "object",
            "properties": {
              "domain": {
                "type": "string",
                "example": "ONDC:FIS13"
              },
              "action": {
                "type": "string",
                "example": "on_select"
              },
              "bap_id": {
                "type": "string",
                "example": "ondc-staging.vaatun.com"
              },
              "bap_uri": {
                "type": "string",
                "example": "https://ondc-staging.vaatun.com/api/ondc"
              },
              "bpp_id": {
                "type": "string",
                "example": "ondc-mock-server.example.com"
              },
              "bpp_uri": {
                "type": "string",
                "example": "https://ondc-mock-server.example.com"
              },
              "transaction_id": {
                "type": "string",
                "example": "019abc12-3456-7890-abcd-ef1234567890"
              },
              "message_id": {
                "type": "string",
                "example": "019abc12-3456-7890-abcd-ef1234567891"
              },
              "timestamp": {
                "type": "string",
                "example": "2024-01-01T00:00:00.000Z"
              },
              "version": {
                "type": "string",
                "example": "2.0.1"
              }
            },
            "additionalProperties": {}
          },
          "message": {
            "type": "object",
            "properties": {
              "order": {
                "type": "object",
                "properties": {
                  "provider": {},
                  "items": {
                    "type": "array",
                    "items": {}
                  },
                  "quote": {
                    "type": "object",
                    "properties": {
                      "price": {
                        "type": "object",
                        "properties": {
                          "currency": {
                            "type": "string",
                            "example": "INR"
                          },
                          "value": {
                            "type": "string",
                            "example": "1000.00"
                          }
                        },
                        "additionalProperties": {}
                      },
                      "breakup": {
                        "type": "array",
                        "items": {}
                      }
                    },
                    "additionalProperties": {}
                  }
                },
                "additionalProperties": {}
              }
            },
            "additionalProperties": {}
          },
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "additionalProperties": {}
          }
        },
        "additionalProperties": {}
      },
      "ProviderSummary": {
        "type": "object",
        "properties": {
          "bppId": {
            "type": "string",
            "example": "ondc-mock-server.example.com"
          },
          "bppUri": {
            "type": "string",
            "example": "https://ondc-mock-server.example.com"
          },
          "name": {
            "type": "string",
            "example": "ICICI Prudential"
          },
          "itemCount": {
            "type": "integer",
            "example": 5
          },
          "hasError": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "bppId",
          "itemCount",
          "hasError"
        ]
      },
      "SearchResultsResponse": {
        "type": "object",
        "properties": {
          "found": {
            "type": "boolean",
            "example": true
          },
          "transactionId": {
            "type": "string",
            "example": "019abc12-3456-7890-abcd-ef1234567890"
          },
          "messageId": {
            "type": "string",
            "example": "019abc12-3456-7890-abcd-ef1234567891"
          },
          "searchTimestamp": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "categoryCode": {
            "type": "string",
            "example": "HEALTH_INSURANCE"
          },
          "responseCount": {
            "type": "integer",
            "example": 3
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderSummary"
            },
            "example": [
              {
                "bppId": "ondc-mock-server.example.com",
                "bppUri": "https://ondc-mock-server.example.com",
                "name": "ICICI Prudential",
                "itemCount": 5,
                "hasError": false
              }
            ]
          },
          "responses": {
            "type": "array",
            "items": {},
            "description": "Raw on_search responses from BPPs"
          },
          "message": {
            "type": "string",
            "example": "No search entry found for this transaction ID"
          }
        },
        "required": [
          "found",
          "transactionId",
          "responseCount",
          "providers",
          "responses"
        ],
        "additionalProperties": {}
      },
...
    },
...
    "securitySchemes": {
      "Ed25519Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "\n# ONDC Ed25519 Signature Authentication\n\nAll outbound requests to ONDC Registry and Gateway require Ed25519 digital signatures using HTTP Signature standard.\n\n## Authentication Flow\n\n### Step 1: Calculate BLAKE-512 Digest\n```typescript\nimport { createHash } from \"crypto\";\n\nconst requestBody = JSON.stringify(payload);\nconst digest = createHash(\"blake2b512\")\n  .update(requestBody)\n  .digest(\"base64\");\n```\n\n### Step 2: Create Signing String\n```typescript\nconst created = Math.floor(Date.now() / 1000) - 3; // Clock drift compensation\nconst expires = created + 300; // 5 minute validity\n\nconst signingString =\n  `(created): ${created}\\n` +\n  `(expires): ${expires}\\n` +\n  `digest: BLAKE-512=${digest}`;\n```\n\n### Step 3: Sign with Ed25519\n```typescript\nimport sodium from \"libsodium-wrappers\";\n\nawait sodium.ready;\n\nconst privateKey = sodium.from_base64(process.env.SIGNING_PRIVATE_KEY);\nconst signature = sodium.crypto_sign_detached(signingString, privateKey);\nconst signatureBase64 = sodium.to_base64(signature);\n```\n\n### Step 4: Construct Authorization Header\n```typescript\nconst keyId = `${subscriberId}|${uniqueKeyId}|ed25519`;\n\nconst authHeader =\n  `Signature keyId=\"${keyId}\",` +\n  `algorithm=\"ed25519\",` +\n  `created=\"${created}\",` +\n  `expires=\"${expires}\",` +\n  `headers=\"(created) (expires) digest\",` +\n  `signature=\"${signatureBase64}\"`;\n```\n\n## Clock Drift Compensation\n\nThe implementation subtracts 3 seconds from the current timestamp to account for clock drift between your server and ONDC servers:\n\n```typescript\nconst created = Math.floor(Date.now() / 1000) - 3; // Subtract 3 seconds\n```\n\nThis prevents authentication failures due to slight time differences.\n\n## Key Management\n\n### Environment Variables Required\n\n- `SIGNING_PRIVATE_KEY`: Ed25519 private key (base64-encoded, 64 bytes)\n- `UNIQUE_KEY_ID`: Unique identifier for this key pair (e.g., \"custom-key-id\")\n- `SUBSCRIBER_ID`: Your registered ONDC subscriber ID (e.g., \"ondc-staging.vaatun.com\")\n\n## Implementation Reference\n\nSee full implementation in:\n- `/src/lib/ondc/client.ts` - ONDCClient.createAuthorizationHeader()\n- `/src/lib/ondc/signing.ts` - Ed25519 signing utilities\n- `/src/entities/tenant.ts` - Key management\n\n## Inbound vs Outbound Authentication\n\n### Outbound (Your App → ONDC)\n- **Routes**: `/api/ondc/lookup`, `/api/ondc/subscribe`, `/api/ondc/search`, `/api/ondc/select`\n- **You sign**: Using your `SIGNING_PRIVATE_KEY`\n- **ONDC verifies**: Using your public key from registry\n\n### Inbound (ONDC → Your App)\n- **Routes**: `/api/ondc/on_subscribe`, `/api/ondc/on_search`, `/api/ondc/on_select`\n- **ONDC signs**: Using their private key\n- **You verify**: Using their public key from registry (retrieved via `/lookup`)\n    "
      }
    }
  },
  "paths": {
...
    "/api/ondc/on_search": {
      "post": {
        "summary": "Receive Search Results",
        "description": "Callback endpoint for receiving search results from BPPs.\n\nThis endpoint is called by BPPs (seller platforms) in response to a search request broadcast by the gateway.\n\n**Flow:**\n1. BPP sends catalog with matching providers and items\n2. Service stores the response keyed by transaction_id\n3. Returns ACK to confirm receipt\n4. Results can be polled via /api/ondc/search-results",
        "tags": [
          "Gateway"
        ],
        "operationId": "onSearch",
        "requestBody": {
          "required": true,
          "description": "Search results from BPP",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/ondc/on_select": {
      "post": {
        "summary": "Receive Quote",
        "description": "Callback endpoint for receiving quotes from BPPs.\n\nThis endpoint is called by BPPs (seller platforms) in response to a select request with pricing details.\n\n**Flow:**\n1. BPP sends order quote with price breakup\n2. Service stores the response keyed by transaction_id and message_id\n3. Returns ACK to confirm receipt\n4. Quote can be polled via /api/ondc/select-results",
        "tags": [
          "Gateway"
        ],
        "operationId": "onSelect",
        "requestBody": {
          "required": true,
          "description": "Quote from BPP",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnSelectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quote acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnSelectAckResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/ondc/search-results": {
      "get": {
        "summary": "Poll Search Results",
        "description": "Polling endpoint to fetch aggregated search results for a transaction.\n\nCalled by the frontend to retrieve on_search responses collected from BPPs.\n\n**Flow:**\n1. Client polls with transaction_id from initial search\n2. Returns all collected on_search responses\n3. Includes provider summary with item counts\n4. Poll until desired providers respond or timeout",
        "tags": [
          "Internal"
        ],
        "operationId": "getSearchResults",
        "directoryConfig": {
          "title": "Poll Search Results",
          "description": "Fetch aggregated search results by transaction ID"
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "019abc12-3456-7890-abcd-ef1234567890",
              "description": "Transaction ID from the initial search request"
            },
            "required": true,
            "description": "Transaction ID from the initial search request",
            "name": "transaction_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results (may be empty if no responses yet)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing transaction_id query parameter"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}